requirements.txt
torch>=2.2
torchvision>=0.17
pillow>=10
pyyaml>=6
tqdm>=4.66
matplotlib>=3.8
scikit-image>=0.22
omegaconf>=2.3